Determine IP address without proprietary DNS servers
The getIPAddress function shouldn't depend on third party proprietary DNS servers that are not part of any Internet standard and may be blocked in some jurisdictions for political reasons.

If the servers behind 8.8.8.8 are inaccessible for any reason, technical or otherwise, the parts of the script that require the IP address to be known won't be functional. It is far more prudent to rely on the 13 root DNS servers that are independent of a company's whim and less likely to be blocked across the world.