joy2key: teach the OSK to start with a value
Added a new parameter to `osk.py` to populate the input box with an existing value.
This makes it easier to replace the `dialog`'s `--inputbox` widget, which accepts an existing value as the last argument.

Removed the prefix from the input prompt, let it be handled by the caller.