audiosettings - use /etc/alsa/conf.d for retropie alsa configs
Move resetting alsa to function to avoid duplication.

Add dialog when migrating ~/.asoundrc to the new location.
If ~/.asoundrc exists and there is no config in /etc/alsa/conf.d/99-retropie.conf then offer to move the file.

Asking the user is safer as we don't want to migrate a users own custom ALSA configuration.