image - set password and disable userconfig in chroot
Set the "pi" password in the chroot install script rather than via userconfig.txt.

Disable the Raspberry Pi OS userconfig service for now. This stops the code to rename user after partition resize.
This would also disable autologin (with a var/lib/userconf-pi/autologin) file, but since our set-up doesn't currently
work with renaming the user RetroPie was installed for due to some hardcoded paths on install, it's safer to bypass this functionality.

Enable getty@tty1.

Note: Have tested with RPI-Imager, which when using a local image, checks it and thinks it can be configured.
This also resolves this case where it would allow a username change.

In the future we may be able to change the few config paths that are hardcoded to allow configuration of the user via RPI-imager.