image - rework logic to use ini configuration for distributions
New function _get_info_image reads values from the distribution ini configuration files stored in image/dists

This simplifies the code and avoids having logic for different versions in multiple functions.

The ini format is as follows - it may change and be extended in the future.
The filename is the dist name that is passed into the building functions.
The current images now have a prefix "rpios-"

 * name: name of distro - eg buster
 * version: rpios/debian version
 * url: url for image file
 * format: format of the image file - currently supported is zip and xz (used by raspberrypi.org)
 * platforms: list of the platforms supported (eg. rpi1 rpi2 rpi3 rpi4)
 * file_PLATFORM (name to be used in the image file for each platform)

 * title_PLATFORM (user friendly name of file for each platform - used for rpi imager json)