runcommand: added option to start a WM for x11 programs
Extend the XINIT: launch prefix with two new ones in order to start a window manager before an x11 application is started.

XINIT-WM: will start a window manager without the cursor
XINIT-WMC: will start a window manager with the cursor

This is similar to how the 'x11' backend starts an application, but without additional configuration set in 'backends.cfg'.