system: fix RPI detection with recent RPI kernels
The 'Hardware' field has been removed from the output of '/proc/cpuinfo' after [1], so the setup script will not correctly identify it's running on a Raspberry PI.
Use '/proc/device-tree/compatible' to check whether we're on a Raspberry Pi, the same method used for other SBCs.

Model calculation based on the Revision field from '/proc/cpuinfo' needs to be run in 2 places, so the code has been put in a separate function.

[1] https://github.com/pelwell/linux/commit/17804c20303c8295b05baf2ab9578432ac5c2735