ports: add 'dsda-doom', lightweight DOOM source port
Added 'dsda-doom', a DOOM source port based on PrBoom+ with extra features geared for demo recording/playback, focused on speedrunning and quality of life.
It has support for:

 * Heretic, Hexen, MBF21, Doom-in-Hexen, UDMF, and MAPINFO support
 * In-game console and scripting
 * Full controller support
 * Palette-based OpenGL renderer
 * Debugging features for testing
 * Strict mode for speedrunning
 * Various quality of life improvements
 * Advanced tools for TASing
 * Rewind

It is meant to replace `lzdoom`, which is no longer developed, as a lightweight DOOM port. Can play most Boom based WADs without issues. Due to its popularity in the TAS community, a lot of BOOM compatible WADs are tested against `dsda-doom`, which should provide better compatiblity than other lightweight source ports (crispy/woof).

It has a SW renderer - which runs ok on a RPI3 - and an OpenGL2 renderer which can be shitched on for OpenGL capable systems. I haven't tested the joypad/gamepad support, though it's there and default mappings take this into account.

Links:
 - Project repository: https://github.com/kraflab/dsda-doom/
 - DoomWorld support topic: https://www.doomworld.com/vb/thread/133096
 - DoomWiki page: https://doomwiki.org/wiki/DSDA-Doom